chore: ExchangeRateDAO refactoring

+++ src/main/java/prog/rohan/currency_conversion/dao/ExchangeRateDAO.java
@@ -4,7 +4,7 @@ import prog.rohan.currency_conversion.model.ExchangeRate;
 
 import java.util.Optional;
 
-public interface ExchangeRateDAO extends BaseDAO<ExchangeRate, Long> {
+public interface ExchangeRateDAO extends BaseDAO<ExchangeRate, Integer> {
 
     Optional<ExchangeRate> findByCodes(String baseCurrencyCode, String targetCurrencyCode);
